Don't make assumptions about layout of content area
I'd had the content area (below the header) showing a span10 and a
span4. I realized that that wasn't going to work for all parts of the
site; settings, especially, needed a different layout.

So, I changed things around a bit to make it easier to have different
layout in the content area.