Add checkRecipient() method to Activity
Add a method checkRecipient() to Activity class, to check if a person
is a recipient of the activity.

Currently it just checks if the person is a direct, named recipient of
the activity; it should check in the future for public activities,
posts to followers, and posts to lists.

Test cases for checkRecipient() added.