Check for bad passwords and nicknames
Check for disallowed nicknames; closes #109.

Also, check for weak passwords. Good passwords are >= 8 chars, aren't
strictly alpha or strictly numeric, and aren't on a list of the 10,000
most used passwords.

We use simple tests for length and chars, and a dictionary to look up
the most-used passwords that are otherwise OK (there are 340, which is
pretty tractable).

I had to do a lot of changes in user-test.js to make the passwords
pass; oops!