Add dispatch module to broker updates between cluster processes
For socket.io, we keep a list of subscribers to a feed in
memory. However, this is distributed across the different clustered
processes, so we need IPC to make it work.

Added a dispatch module that takes follow, unfollow, and update
commands mapped to an URL. When it gets an update, it sends a message
to each child process that has followed the URL. The child then
distributes to each of its sockets that is following the url.

Probably going to be a mess, but I'm excited to find out.