Massive breakup of pump.js
pump.js was flirting with 3000 lines and was becoming unusable.

It's now broken into a half-dozen modules. One (view.js) is still 1.8K
lines, so there's still some modularization that needs to be done. But
it's a start.

Ultimately there should probably be a module per class, like
pump/view/inbox.js and pump/model/activityobjectbag.js. Another
possibility is clustering around sections of the site.