Change big switch-case in apply() to separate applyVerb() methods
I had a big switch-case statement in apply() that would do different
things depending on the verb.

It now tries to find a method of the Activity prototype based on a
camel-cased version of the verb. So like => applyLike(),
stop-following => applyStopFollowing().

If the verb is unrecognized, or if there's no method for the verb, it
continues silently.