Switch remoteUser to principalUser
This is kind of a sweeping change to support having identified
principals (= person who is using the site) who are not users (people
who have an account on the site).

It entails changing the authentication methods to set different
properties on the request, and checking those properties in the API code.

It's mostly working, but there seems to be some problem with the
finishers which I'm going to fix right away.