More informative error message on registration
This commit changes the register page template to add more informative
error messages, as obtained from the server. If the jqXHR is text/plain,
dislay its responseText, if it is text/html, try to extract the title
for display, or just the entire page otherwise.

This should probably be generalised to the rest of the templates. It
might be possible to create micro templates doing that once and for all.

refs: #623

Signed-off-by: Olivier Mehani <shtrom@ssji.net>