Submodule the openshift dir to enable deployment
This seems to be the cleanest way to add the .openshift directory based
on the grdryn/nodejs-custom-version-openshift repository, which is a
customized fork of the openshift/nodejs-custom-version-openshift repo.

This will enable us to add any changes from the upstream
openshift/nodejs-custom-version-openshift repo, keeping traceability,
while keeping the pump.io repository clean.

So, the good stuff is that with this change, one can deploy pump.io to
openshift as simply as (assuming rhc is installed):

$ rhc app create <mypump> nodejs-0.6 mongodb-2.2 \
	--from-code=git://github.com/e14n/pump.io.git

...or the equivalent from the web interface.