Replaced bro image with picture of community members
The photograph is cc-by-sa, but I have waived it as CC0 to make things
simple for including it in pump-io.  The photo is of the MediaGoblin
OPW volunteers (participants and a mentor) giving Bassam Kurdali a
group-hug.  Jessica Tallon, seen on the left is currently adding
federation support via the pump API to MediaGoblin.