Show a message if Javascript is disabled
If you have Javascript disabled, the login and register pages now post
to a page that explains that Javascript is needed to run this Web app.

Previously, the forms had no "action", because they use Javascript,
but if Javascript is disabled they would act as a GET form by default,
showing the username and password in the URL bar and in server logs.