Send the full hostname in /api/whoami redirects
This changed with Express 3.x, so we restore the old behavior.

In particular this broke whoami-test.js in a _very_ subtle way: the
OAuth module that test uses will follow redirects, but without this
change, it would simply get a relative path back. Since that path didn't
include the port, it would default to localhost port 80 and give a
_very_ cryptic ECONNREFUSED error.

The above seems rather like a bug in said OAuth library, but frankly
this API worked this way already and this seems much easier than going
and patching upstream. (Note also that upstream in this case is an E14N
fork, and I don't have time to verify if this was fixed in the original
module, etc.)