Split out Jade that the client needs to use
The problem here is that all the Jade extends from layout.jade as of the
Express 3.x merge. Because of this the Jade that the client gets is for
the _entire_ page, but the client is expecting only one particular part.

Therefore, we split the client parts out into "*-client.jade" files and
request those. The old Jade files still inherit from layout and simply
include the client files.