Use system uid, prep & change data & install dirs
This patch changes the UID and GUID for the pumpio system user into
the system range (< 1000).

The installation location for the pump.io source is moved to a local
dir instead of the npm installation directory.

The datadir directory is prepared and permissions are set so that
uploads will work out of the box. For persistence users should use a
volume for this location.