Set req.pumphost instead of req.host
Express 3.x added the `req.host` property which has only a getter, not
a setter. This meant that our usage of it conflicted with that and, to
make matters worse, the error wouldn't be caught by Express's normal
500 middleware because it was being called in a process.nextTick
handler due to some weird stuff Step was doing.

In any case, we only use this internally and don't really care about
the new Express thing, so just rename the property so it doesn't
clash (and crash).

Fixes #1281