Fix the null string being piped to `coveralls`
The problem here is that `lcov-result-merger` expects exactly one or
exactly two arguments; the first is expected to be a glob which will
be expanded _internally in the program_ and the second is expected to
be an output file. Because of this, before this change,
`lcov-result-merger` would intepret the as-root coverage data as the
output file and write data there instead of piping it to
stdout. `coveralls` would thus receive no data from stdin and
interpret this as a null string, causing the "cannot parse lcov" error in
Travis.

So we adjust the invocation so that we only use one argument, a glob
which matches both sets of coverage data. Since it doesn't have a
second argument `lcov-result-merger` will start behaving as one might
expect and output to stdout, which then gets piped to Coveralls.