Switch to connect-auth-pumpio
We made this change long ago on `master`, but we have to do it on this
maintenance branch in order to pull in an updated Connect, fixing
denial-of-service security vulnerabilities.

This is basically the same code so this isn't actually all that
risky. Additionally this release will be delayed until CI runs, which
will give us some confidence it isn't introducing instability.