Fix worker.suicide deprecation warnings
The problem was that || evaluates the right side not just if the left
side returns `undefined` (which is what we were looking for), but in
fact if the left side returns _any_ falsy value including, well,
`false`.

This is obviously the right behavior and a Good Thing™, and I was
being dumb in this case. So fix the logic to actually properly check
for what we're looking for.