s/JSON.parse(fs.readFileSync())/require()/ in test/
`require()` will also load JSON files in addition to `.js` files. We
take advantage of this to simplify the logic here. The other big
advantage is that `require()` caches results, which will eliminate a
lot of slow disk I/O in the test suite when multiple tests are run in
the same process.

We surround the `require()` with `_.clone()` because the tests tend to
fiddle with the config we pull here, and that might cause problems
with the `require()` caching.