Reduce the amount of code we load as root
Fixes #1555. This is particularly important since Node's
require() has to actively execute each module required, and we
don't want just any module in the dep tree to be able to run
arbitrary code as root on the administrator's system.

As a bonus, this improves startup performance; `time node -r
./lib/app -e true` reports 3.341s total whereas `time node -e
true` reports only 0.092s total. That means we shaved off a
little under three and a half seconds from startup.